Fixed wrap around w/ stack pointer variable
- Now using a fixed 8-bit (uint8_t) variable to hold the offset to the
  stack pointer (which automatically wraps around)
- cNES now also passes Blargg's offical opcodes test from instr_test-v5

+++ src/extern_structs.h
@@ -131,7 +131,7 @@ typedef struct {
 	/* Special Registers */
 	uint8_t P; /* Program status register */
 	uint16_t PC; /* Program counter (instruction pointer) */
-	int stack;
+	uint8_t stack;
 	unsigned cycle; /* Helper variable, logs how many cpu cycles have elapsed */
 	/* Memory */
 	uint8_t mem[CPU_MEMORY_SIZE];
